+++ src/timer.rs
@@ -7,6 +7,7 @@
 use core::convert::TryFrom;
 use cortex_m::peripheral::syst::SystClkSource;
 use cortex_m::peripheral::SYST;
+use enumflags2::BitFlags;
 
 use crate::bb;
 use crate::pac;
@@ -79,7 +80,7 @@ pub enum IdleState {
     Set,
 }
 
-/// Interrupt events
+/// SysTick interrupt events
 #[derive(Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum SysEvent {
@@ -87,14 +88,60 @@ pub enum SysEvent {
     Update,
 }
 
-bitflags::bitflags! {
-    pub struct Event: u32 {
-        const Update  = 1 << 0;
-        const C1 = 1 << 1;
-        const C2 = 1 << 2;
-        const C3 = 1 << 3;
-        const C4 = 1 << 4;
-    }
+/// TIM interrupt events
+#[enumflags2::bitflags]
+#[repr(u8)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Event {
+    /// Update interrupt enable
+    Update = 1 << 0,
+    /// Capture/Compare 1 interrupt enable
+    C1 = 1 << 1,
+    /// Capture/Compare 2 interrupt enable
+    C2 = 1 << 2,
+    /// Capture/Compare 3 interrupt enable
+    C3 = 1 << 3,
+    /// Capture/Compare 4 interrupt enable
+    C4 = 1 << 4,
+    /// COM interrupt enable
+    COM = 1 << 5,
+    /// Trigger interrupt enable
+    Trigger = 1 << 6,
+    /// Break interrupt enable
+    Break = 1 << 7,
+}
+
+/// TIM status flags
+#[enumflags2::bitflags]
+#[repr(u16)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Flag {
+    /// Update interrupt flag
+    Update = 1 << 0,
+    /// Capture/Compare 1 interrupt flag
+    C1 = 1 << 1,
+    /// Capture/Compare 2 interrupt flag
+    C2 = 1 << 2,
+    /// Capture/Compare 3 interrupt flag
+    C3 = 1 << 3,
+    /// Capture/Compare 4 interrupt flag
+    C4 = 1 << 4,
+    /// COM interrupt flag
+    COM = 1 << 5,
+    /// Trigger interrupt flag
+    Trigger = 1 << 6,
+    /// Break interrupt flag
+    Break = 1 << 7,
+    /// Capture/Compare 1 overcapture flag
+    C1Overcapture = 1 << 9,
+    /// Capture/Compare 2 overcapture flag
+    C2Overcapture = 1 << 10,
+    /// Capture/Compare 3 overcapture flag
+    C3Overcapture = 1 << 11,
+    /// Capture/Compare 4 overcapture flag
+    C4Overcapture = 1 << 12,
 }
 
 #[derive(Debug, Eq, PartialEq, Copy, Clone)]
@@ -254,7 +301,7 @@ pub type CCR4<T> = CCR<T, 3>;
 pub struct DMAR<T>(T);
 
 mod sealed {
-    use super::{Channel, Event, IdleState, Ocm, Polarity};
+    use super::{BitFlags, Channel, Event, Flag, IdleState, Ocm, Polarity};
     pub trait General {
         type Width: Into<u32> + From<u16>;
         fn max_auto_reload() -> u32;
@@ -262,16 +309,19 @@ mod sealed {
         fn set_auto_reload(&mut self, arr: u32) -> Result<(), super::Error>;
         fn read_auto_reload() -> u32;
         fn enable_preload(&mut self, b: bool);
-        fn enable_counter(&mut self);
-        fn disable_counter(&mut self);
+        fn enable_counter(&mut self, b: bool);
         fn is_counter_enabled(&self) -> bool;
         fn reset_counter(&mut self);
         fn set_prescaler(&mut self, psc: u16);
         fn read_prescaler(&self) -> u16;
         fn trigger_update(&mut self);
-        fn clear_interrupt_flag(&mut self, event: Event);
-        fn listen_interrupt(&mut self, event: Event, b: bool);
-        fn get_interrupt_flag(&self) -> Event;
+        fn listen_event(
+            &mut self,
+            disable: Option<BitFlags<Event>>,
+            enable: Option<BitFlags<Event>>,
+        );
+        fn clear_interrupt_flag(&mut self, event: BitFlags<Flag>);
+        fn get_interrupt_flag(&self) -> BitFlags<Flag>;
         fn read_count(&self) -> Self::Width;
         fn write_count(&mut self, value: Self::Width);
         fn start_one_pulse(&mut self);
@@ -356,12 +406,8 @@ macro_rules! hal {
                 self.cr1.modify(|_, w| w.arpe().bit(b));
             }
             #[inline(always)]
-            fn enable_counter(&mut self) {
-                self.cr1.modify(|_, w| w.cen().set_bit());
-            }
-            #[inline(always)]
-            fn disable_counter(&mut self) {
-                self.cr1.modify(|_, w| w.cen().clear_bit());
+            fn enable_counter(&mut self, b: bool) {
+                self.cr1.modify(|_, w| w.cen().bit(b));
             }
             #[inline(always)]
             fn is_counter_enabled(&self) -> bool {
@@ -386,20 +432,25 @@ macro_rules! hal {
                 self.cr1.modify(|_, w| w.urs().clear_bit());
             }
             #[inline(always)]
-            fn clear_interrupt_flag(&mut self, event: Event) {
-                self.sr.write(|w| unsafe { w.bits(0xffff & !event.bits()) });
+            fn listen_event(&mut self, disable: Option<BitFlags<Event>>, enable: Option<BitFlags<Event>>) {
+                self.dier.modify(|r, w| unsafe { w.bits({
+                    let mut bits = r.bits();
+                    if let Some(d) = disable {
+                        bits &= !(d.bits() as u32);
+                    }
+                    if let Some(e) = enable {
+                        bits |= e.bits() as u32;
+                    }
+                    bits
+                }) });
             }
             #[inline(always)]
-            fn listen_interrupt(&mut self, event: Event, b: bool) {
-                if b {
-                    self.dier.modify(|r, w| unsafe { w.bits(r.bits() | event.bits()) });
-                } else {
-                    self.dier.modify(|r, w| unsafe { w.bits(r.bits() & !event.bits()) });
-                }
+            fn clear_interrupt_flag(&mut self, event: BitFlags<Flag>) {
+                self.sr.write(|w| unsafe { w.bits(0xffff & !(event.bits() as u32)) });
             }
             #[inline(always)]
-            fn get_interrupt_flag(&self) -> Event {
-                Event::from_bits_truncate(self.sr.read().bits())
+            fn get_interrupt_flag(&self) -> BitFlags<Flag> {
+                BitFlags::from_bits_truncate(self.sr.read().bits() as u16)
             }
             #[inline(always)]
             fn read_count(&self) -> Self::Width {
@@ -617,27 +668,6 @@ impl<TIM: Instance> Timer<TIM> {
     pub fn release(self) -> TIM {
         self.tim
     }
-
-    /// Starts listening for an `event`
-    ///
-    /// Note, you will also have to enable the TIM2 interrupt in the NVIC to start
-    /// receiving events.
-    pub fn listen(&mut self, event: Event) {
-        self.tim.listen_interrupt(event, true);
-    }
-
-    /// Clears interrupt associated with `event`.
-    ///
-    /// If the interrupt is not cleared, it will immediately retrigger after
-    /// the ISR has finished.
-    pub fn clear_interrupt(&mut self, event: Event) {
-        self.tim.clear_interrupt_flag(event);
-    }
-
-    /// Stops listening for an `event`
-    pub fn unlisten(&mut self, event: Event) {
-        self.tim.listen_interrupt(event, false);
-    }
 }
 
 impl<TIM: Instance + MasterTimer> Timer<TIM> {
@@ -697,31 +727,6 @@ impl<TIM: Instance, const FREQ: u32> FTimer<TIM, FREQ> {
     pub fn release(self) -> TIM {
         self.tim
     }
-
-    /// Starts listening for an `event`
-    ///
-    /// Note, you will also have to enable the TIM2 interrupt in the NVIC to start
-    /// receiving events.
-    pub fn listen(&mut self, event: Event) {
-        self.tim.listen_interrupt(event, true);
-    }
-
-    /// Clears interrupt associated with `event`.
-    ///
-    /// If the interrupt is not cleared, it will immediately retrigger after
-    /// the ISR has finished.
-    pub fn clear_interrupt(&mut self, event: Event) {
-        self.tim.clear_interrupt_flag(event);
-    }
-
-    pub fn get_interrupt(&self) -> Event {
-        self.tim.get_interrupt_flag()
-    }
-
-    /// Stops listening for an `event`
-    pub fn unlisten(&mut self, event: Event) {
-        self.tim.listen_interrupt(event, false);
-    }
 }
 
 impl<TIM: Instance + MasterTimer, const FREQ: u32> FTimer<TIM, FREQ> {
@@ -736,6 +741,62 @@ pub(crate) const fn compute_arr_presc(freq: u32, clock: u32) -> (u16, u32) {
     let psc = (ticks - 1) / (1 << 16);
     let arr = ticks / (psc + 1) - 1;
     (psc as u16, arr)
+}
+
+impl<TIM: Instance> crate::Listen for Timer<TIM> {
+    type Event = Event;
+    fn listen(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim.listen_event(None, Some(event.into()));
+    }
+    fn listen_only(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim
+            .listen_event(Some(BitFlags::ALL), Some(event.into()));
+    }
+    fn unlisten(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim.listen_event(Some(event.into()), None);
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> crate::Listen for FTimer<TIM, FREQ> {
+    type Event = Event;
+    fn listen(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim.listen_event(None, Some(event.into()));
+    }
+    fn listen_only(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim
+            .listen_event(Some(BitFlags::ALL), Some(event.into()));
+    }
+    fn unlisten(&mut self, event: impl Into<BitFlags<Event>>) {
+        self.tim.listen_event(Some(event.into()), None);
+    }
+}
+
+impl<TIM: Instance> crate::ClearFlags for Timer<TIM> {
+    type Flag = Flag;
+    fn clear_flags(&mut self, event: impl Into<BitFlags<Flag>>) {
+        self.tim.clear_interrupt_flag(event.into());
+    }
+}
+
+impl<TIM: Instance> crate::ReadFlags for Timer<TIM> {
+    type Flag = Flag;
+    fn flags(&self) -> BitFlags<Flag> {
+        self.tim.get_interrupt_flag()
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> crate::ClearFlags for FTimer<TIM, FREQ> {
+    type Flag = Flag;
+    fn clear_flags(&mut self, event: impl Into<BitFlags<Flag>>) {
+        self.tim.clear_interrupt_flag(event.into());
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> crate::ReadFlags for FTimer<TIM, FREQ> {
+    type Flag = Flag;
+    fn flags(&self) -> BitFlags<Flag> {
+        self.tim.get_interrupt_flag()
+    }
 }
 
 #[cfg(not(feature = "gpio-f410"))]